Develop

+++ yo/app/scripts/controllers/jobsCtrl.js
@@ -6,41 +6,39 @@
 'use strict';
 
 angular.module('hopsWorksApp')
-        .controller('JobsCtrl', ['$scope', '$routeParams', 'growl', 'JobService', '$location', 'ModalService', '$interval','StorageService','$mdSidenav',
-          function ($scope, $routeParams, growl, JobService, $location, ModalService, $interval,StorageService,$mdSidenav) {
+        .controller('JobsCtrl', ['$scope', '$routeParams', 'growl', 'JobService', '$location', 'ModalService', '$interval', 'StorageService',
+          '$mdSidenav', 'TourService', 'ProjectService',
+          function ($scope, $routeParams, growl, JobService, $location, ModalService, $interval, StorageService,
+                  $mdSidenav, TourService, ProjectService) {
 
             var self = this;
-            this.projectId = $routeParams.projectID;
-            this.jobs; // Will contain all the jobs.
-            this.runningInfo; //Will contain run information
-            this.buttonArray = [];
-            this.workingArray = [];
-            this.jobFilter = {
-              "creator":{
-                "email":""
-              },
-              "jobType":"",
-              "name":""
-            };
-            
-            this.hasSelectJob=false;
-            
-            self.currentjob=null;
-            self.currentToggledIndex=-1;
+            self.tourService = TourService;
+            self.projectId = $routeParams.projectID;
+            self.jobs; // Will contain all the jobs.
+            self.runningInfo; //Will contain run information
+            self.buttonArray = [];
+            self.workingArray = [];
+            self.jobFilter = "";
+
+            self.hasSelectJob = false;
+
+            self.currentjob = null;
+            self.currentToggledIndex = -1;
 
             $scope.pageSize = 10;
             $scope.sortKey = 'creationTime';
             $scope.reverse = true;
 
-            $scope.sort = function(keyname){
+            $scope.sort = function (keyname) {
               $scope.sortKey = keyname;   //set the sortKey to the param passed
               $scope.reverse = !$scope.reverse; //if true make it false and vice versa
-            }
-            
-            this. editAsNew = function (job) {   
+            };
+
+
+            self.editAsNew = function (job) {
               JobService.getConfiguration(self.projectId, job.id).then(
                       function (success) {
-                        self.currentjob=job;
+                        self.currentjob = job;
                         self.currentjob.runConfig = success.data;
                         self.copy();
                       }, function (error) {
@@ -48,40 +46,35 @@ angular.module('hopsWorksApp')
               });
             };
 
-            self.buttonClickedToggle = function (id, display){
+            self.buttonClickedToggle = function (id, display) {
               self.buttonArray[id] = display;
-            }
+            };
 
-            self.stopbuttonClickedToggle = function (id, display){
+            self.stopbuttonClickedToggle = function (id, display) {
               self.workingArray[id] = display;
-            }
-            
+            };
+
             self.copy = function () {
               var jobType;
               switch (self.currentjob.jobType.toUpperCase()) {
-                case "CUNEIFORM":
-                  jobType = 0;
-                  break;
                 case "SPARK":
                   jobType = 1;
                   break;
                 case "ADAM":
                   jobType = 2;
                   break;
+                case "FLINK":
+                  jobType = 3;
               }
-              var mainFileTxt, mainFileVal, jobDetailsTxt, sparkState, adamState;
-              if (jobType == 0) {
-                mainFileTxt = "Workflow file";
-                mainFileVal = self.currentjob.runConfig.wf.name;
-                jobDetailsTxt = "Input variables";
-              } else if (jobType == 1) {
+              var mainFileTxt, mainFileVal, jobDetailsTxt, sparkState, adamState, flinkState;
+              if (jobType === 1) {
                 sparkState = {
                   "selectedJar": getFileName(self.currentjob.runConfig.jarPath)
                 };
                 mainFileTxt = "JAR file";
                 mainFileVal = sparkState.selectedJar;
                 jobDetailsTxt = "Job details";
-              } else if (jobType == 2) {
+              } else if (jobType === 2) {
                 adamState = {
                   "processparameter": null,
                   "commandList": null,
@@ -90,6 +83,13 @@ angular.module('hopsWorksApp')
                 mainFileTxt = "ADAM command";
                 mainFileVal = adamState.selectedCommand;
                 jobDetailsTxt = "Job arguments";
+              } else if (jobType === 3) {
+                flinkState = {
+                  "selectedJar": getFileName(self.currentjob.runConfig.jarPath)
+                };
+                mainFileTxt = "JAR file";
+                mainFileVal = flinkState.selectedJar;
+                jobDetailsTxt = "Job details";
               }
               var state = {
                 "jobtype": jobType,
@@ -99,6 +99,7 @@ angular.module('hopsWorksApp')
                 "runConfig": self.currentjob.runConfig,
                 "sparkState": sparkState,
                 "adamState": adamState,
+                "flinkState": flinkState,
                 "accordion1": {//Contains the job name
                   "isOpen": false,
                   "visible": true,
@@ -123,32 +124,52 @@ angular.module('hopsWorksApp')
                   "isOpen": false,
                   "visible": true,
                   "value": "",
-                  "title": "Configure and create"}
+                  "title": "Configure and create"},
+                "accordion6" : {//Contains the pre-configuration and proposals for auto-configuration
+                   "isOpen": false,
+                   "visible": true,
+                   "value": "",
+                   "title": "Pre-Configuration"}
               };
-              StorageService.store(self.projectId + "newjob", state);              
+              StorageService.store(self.projectId + "_newjob", state);
               $location.path('project/' + self.projectId + '/newjob');
             };
-            
-            
+
+
 
             var getAllJobs = function () {
               JobService.getAllJobsInProject(self.projectId).then(
                       function (success) {
                         self.jobs = success.data;
                         angular.forEach(self.jobs, function (job, key) {
-                        job.showing = false;
-                      });
+                          job.showing = false;
+                        });
                       }, function (error) {
                 growl.error(error.data.errorMsg, {title: 'Error', ttl: 15000});
               });
             };
+            
+            self.getNumOfExecution = function () {
+              if (self.hasSelectJob) {
+                if (self.logset === undefined) {
+                  return 0;
+                }
+                if (self.logset.length > 1) {
+                  return self.logset.length;
+                } else if (self.logset.length === 1 && self.logset[0].appId !== '') {
+                  return 1;
+                } else {
+                  return 0;
+                }
+              }
+            };
 
-            this.getRunStatus = function () {
+            self.getRunStatus = function () {
               JobService.getRunStatus(self.projectId).then(
                       function (success) {
                         self.runningInfo = success.data;
                         angular.forEach(self.jobs, function (temp, key) {
-                          if (typeof self.runningInfo['' + temp.id] !== "undefined"){
+                          if (typeof self.runningInfo['' + temp.id] !== "undefined") {
                             if (!self.runningInfo['' + temp.id].running) {
                               self.buttonArray[temp.id] = false;
                             }
@@ -159,10 +180,10 @@ angular.module('hopsWorksApp')
               });
             };
 
-            this.createAppReport = function () {
+            self.createAppReport = function () {
               angular.forEach(self.jobs, function (temp, key) {
-                if (typeof self.runningInfo['' + temp.id] !== "undefined"){
-                  if (temp.state !== self.runningInfo['' + temp.id].state){
+                if (typeof self.runningInfo['' + temp.id] !== "undefined") {
+                  if (temp.state !== self.runningInfo['' + temp.id].state) {
                     self.showLogs(temp.id);
                   }
                   temp.duration = self.runningInfo['' + temp.id].duration;
@@ -171,32 +192,50 @@ angular.module('hopsWorksApp')
                   temp.running = self.runningInfo['' + temp.id].running;
                   temp.state = self.runningInfo['' + temp.id].state;
                   temp.submissiontime = self.runningInfo['' + temp.id].submissiontime;
+                  temp.url = self.runningInfo['' + temp.id].url;
                 }
-              })
+              });
             };
-
             getAllJobs();
             self.getRunStatus();
             self.createAppReport();
 
-            this.runJob = function (jobId) {
-              JobService.runJob(self.projectId, jobId).then(
+            self.runJob = function (job, index) {
+              var jobId = job.id;
+
+              ProjectService.uberPrice({id: self.projectId}).$promise.then(
                       function (success) {
-                        self.getRunStatus();
+                        var price = success.price;
+                        price = Math.ceil(parseFloat(price).toFixed(4) * 100.0 / 1.67 * 100) / 100;
+                        ModalService.uberPrice('sm', 'Confirm', 'Do you want to run this job at this price?', price).then(
+                                function (success) {
+                                  JobService.runJob(self.projectId, jobId).then(
+                                          function (success) {
+                                            self.toggle(job, index);
+                                            self.buttonClickedToggle(job.id, true);
+//                                            self.stopbuttonClickedToggle(job.id, false);
+                                            self.getRunStatus();
+                                          }, function (error) {
+                                    growl.error(error.data.errorMsg, {title: 'Failed to run job', ttl: 10000});
+                                  });
+
+                                }
+                        );
+
                       }, function (error) {
-                growl.error(error.data.errorMsg, {title: 'Failed to run job', ttl: 15000});
+                growl.error(error.data.errorMsg, {title: 'Could not get the current YARN price.', ttl: 10000});
               });
             };
 
-            this.stopJob = function (jobId) {
+            self.stopJob = function (jobId) {
               self.stopbuttonClickedToggle(jobId, true);
               JobService.stopJob(self.projectId, jobId).then(
-                function (success) {
-                  self.getRunStatus();
-                }, function (error) {
-                  growl.error(error.data.errorMsg, {title: 'Failed to stop' +
-                  ' job', ttl: 15000});
-                });
+                      function (success) {
+                        self.getRunStatus();
+                      }, function (error) {
+                growl.error(error.data.errorMsg, {title: 'Failed to stop' +
+                          ' job', ttl: 15000});
+              });
             };
 
             /**
@@ -206,38 +245,60 @@ angular.module('hopsWorksApp')
             self.newJob = function () {
               StorageService.clear();
               $location.path('project/' + self.projectId + '/newjob');
+              if (self.tourService.currentStep_TourThree > -1) {
+                self.tourService.resetTours();
+              }
             };
 
             self.showDetails = function (job) {
               ModalService.jobDetails('lg', job, self.projectId);
             };
-            
-            
+
+            self.showUI = function (job) {
+              ModalService.jobUI('xlg', job, self.projectId);
+            };
+
             self.showLogs = function (jobId) {
-              JobService.showLog(self.projectId,jobId).then(
-                      function (success) {                          
-                          self.logset=success.data.logset;                         
+              JobService.showLog(self.projectId, jobId).then(
+                      function (success) {
+                        self.logset = success.data.logset;
                       }, function (error) {
                 growl.error(error.data.errorMsg, {title: 'Failed to show logs', ttl: 15000});
-              });              
-            };
-            
-            self.deleteJob=function (jobId){
-
-                  JobService.deleteJob(self.projectId,jobId).then(
-                      function (success) {
-                           getAllJobs();
-                           self.hasSelectJob=false;
-                           growl.success(success.data.successMessage, {title: 'Success', ttl: 5000});
-                      }, function (error) {
-                growl.error(error.data.errorMsg, {title: 'Failed to delete job', ttl: 15000});
               });
-
- 
             };
-            
-           
-            self.toggle = function (job,index) {
+
+            self.retryLogs = function (appId, type) {
+              if (appId === '' || appId === undefined) {
+                growl.error("Can not retry log. The job has not yet been assigned an Id", {title: 'Error', ttl: 5000});
+              }
+              JobService.retryLog(self.projectId, appId, type).then(
+                      function (success) {
+                        growl.success(success.data.successMessage, {title: 'Success', ttl: 5000});
+                        self.showLogs(self.currentjob.id);
+                      }, function (error) {
+                        growl.error(error.data.errorMsg, {title: 'Failed to get logs', ttl: 5000});
+              });
+            };
+
+            self.deleteJob = function (jobId, jobName) {
+              ModalService.confirm("sm", "Delete Job (" + jobName + ")",
+                      "Do you really want to delete this job?\n\
+                                This action cannot be undone.")
+                      .then(function (success) {
+                        JobService.deleteJob(self.projectId, jobId).then(
+                                function (success) {
+                                  getAllJobs();
+                                  self.hasSelectJob = false;
+                                  growl.success(success.data.successMessage, {title: 'Success', ttl: 5000});
+                                }, function (error) {
+                          growl.error(error.data.errorMsg, {title: 'Failed to delete job', ttl: 15000});
+                        });
+                      }, function (cancelled) {
+                        growl.info("Delete aborted", {title: 'Info', ttl: 5000});
+                      });
+            };
+
+            self.toggle = function (job, index) {
               //reset all jobs showing flag
               angular.forEach(self.jobs, function (temp, key) {
                 if (job.id !== temp.id) {
@@ -247,35 +308,40 @@ angular.module('hopsWorksApp')
 
               //handle the clicked job accordingly
               job.showing = true;
-              self.hasSelectJob=true;
-              $scope.selectedIndex=index;
-              self.currentToggledIndex=index;
-
+              self.hasSelectJob = true;
+              $scope.selectedIndex = index;
+              self.currentToggledIndex = index;
+              self.currentjob = job;
             };
-            self.untoggle = function (job,index) {
+            self.untoggle = function (job, index) {
               //reset all jobs showing flag
               angular.forEach(self.jobs, function (temp, key) {
-                   temp.showing = false;
+                temp.showing = false;
               });
-              
-              if(self.currentToggledIndex !== index){
-                  self.hasSelectJob=false;
-                  $scope.selectedIndex=-1;
-                  self.currentToggledIndex=-1;
-              }else{
-                  job.showing = true;
+
+              if (self.currentToggledIndex !== index) {
+                self.hasSelectJob = false;
+                $scope.selectedIndex = -1;
+                self.currentToggledIndex = -1;
+              } else {
+                job.showing = true;
               }
             };
-            
+
             /**
              * Check if the jobType filter is null, and set to empty string if it is.
              * @returns {undefined}
              */
-            this.checkJobTypeFilter = function(){
-              if(self.jobFilter.jobType == null){
+            self.checkJobTypeFilter = function () {
+              if (self.jobFilter.jobType === null) {
                 self.jobFilter.jobType = "";
               }
             };
+            
+            self.launchAppMasterUrl = function (trackingUrl) {
+              window.open(trackingUrl);
+            };
+
 
             /**
              * Close the poller if the controller is destroyed.
@@ -292,6 +358,34 @@ angular.module('hopsWorksApp')
             };
             startPolling();
 
+            $scope.convertMS = function (ms) {
+              if (ms === undefined) {
+                return "";
+              }
+              var m, s;
+              s = Math.floor(ms / 1000);
+              m = Math.floor(s / 60);
+              s = s % 60;
+              if (s.toString().length < 2) {
+                s = '0' + s;
+              }
+              if (m.toString().length < 2) {
+                m = '0' + m;
+              }
+              var ret = m + ":" + s;
+              return ret;
+            };
+
+
+            var init = function () {
+              var stored = StorageService.contains(self.projectId + "_newjob");
+              if (stored) {
+//                self.newJob();
+                  $location.path('project/' + self.projectId + '/newjob');
+              }
+            };
+
+            init();
           }]);
